Fix a linking error affecting GCC on Ubuntu
Blindly following the examples in a tutorial, I used the variable
AM_LDFLAGS to pass the required libraries instead of program_LDADD.
This worked when using Clang on OpenBSD 7.2, but on Ubuntu using
the GCC, it passes the arguments to the GNU linker in the wrong
order. This resulted in a lot of linker errors.

After switching to the correct Autotools variable, the program
builds successfully on Ubuntu using the GCC.