Ensure string is terminated by nul character
Shader source code is read from a seperate file into a string. The
function load_shader_src() does not add a NUL character at the end
of that string.

At random, OpenGL fails to compile shaders, and complains of
unexpected characters encountered in the source string. This is
consistent with glShaderSource() reading past the end of a string
that is not terminated by a NUL character.

This commit ensures that the string containing GLSL source code
will always be terminated by a NUL character.